aco

Struct Ant

pub(crate) struct Ant<'a> {
    pub(crate) aco_config: &'a ACOConfig,
    pub(crate) problem: &'a Problem,
    pub(crate) tabu_list: Vec<bool>,
    pub(crate) curr_pos: usize,
    pub(crate) tour: Vec<usize>,
}

Representa uma formiga no algoritmo ACO com os atributos necessários para sua operação e tomada de decisões

Parameters

aco_config - Referência à configuração do ACO

problem - Referência à instância do problema

tabu_list - Lista tabu para rastrear tarefas já visitadas

curr_pos - Posição atual da formiga

tour - Sequência de tarefas visitadas pela formiga

Fields
aco_config: &'a ACOConfigproblem: &'a Problemtabu_list: Vec<bool>curr_pos: usizetour: Vec<usize>
Implementations

impl<'a> Ant<'a>

pub(crate) fn init( aco_config: &'a ACOConfig, problem: &'a Problem, start_pos: usize, ) -> Self

Construtor para criar uma nova formiga com os parâmetros fornecidos Retorna uma nova formiga

Arguments

aco_config - Referência à configuração do ACO

problem - Referência à instância do problema

start_pos - Posição inicial da formiga

Returns

Self - Nova formiga posicionada na tarefa inicial

pub(crate) fn move_to(&mut self, next_pos: usize)

Move a formiga para a próxima posição

Atualiza a posição atual da formiga

Marca a posição como visitada na lista tabu

Adiciona a posição ao tour da formiga

Arguments

next_pos - Próxima posição para a qual a formiga deve se mover

pub(crate) fn choose_next_position( &self, pheromone_matrix: &mut Vec<Vec<f64>>, ) -> Option<usize>

Escolhe a próxima posição para a formiga se mover com base na matriz de feromônio e nas regras do ACO

Arguments

pheromone_matrix - Matriz de feromônio utilizada para a decisão

Returns

Option<usize> - Próxima posição escolhida ou None se não houver