aco

Struct Problem

pub(crate) struct Problem {
    pub(crate) n: usize,
    pub(crate) processing_time: Vec<f64>,
    pub(crate) weight: Vec<f64>,
    pub(crate) due_date: Vec<f64>,
}

Representa uma instância do problema de sequenciamento de tarefas

Fields
n: usize

Número de tarefas

processing_time: Vec<f64>

Tempo de processamento de cada tarefa

weight: Vec<f64>

Peso de cada tarefa

due_date: Vec<f64>

Data de vencimento de cada tarefa

Implementations

impl Problem

pub(crate) fn new( n: usize, processing_time: Vec<f64>, weight: Vec<f64>, due_date: Vec<f64>, ) -> Self

Construtor para criar uma nova instância do problema com os parâmetros fornecidos Retorna uma nova instância do problema

Arguments

n - Número de tarefas

processing_time - Vetor com o tempo de processamento de cada tarefa

weight - Vetor com o peso de cada tarefa

due_date - Vetor com a data de vencimento de cada tarefa

Returns

Self - Nova instância do problema

pub(crate) fn get_toy_problem() -> Self

Função para obter um problema de brinquedo (toy problem) Retorna uma nova instância do problema de brinquedo

Returns

Self - Nova instância do problema de brinquedo Dados: | Tarefa | Tempo de Processamento | Peso | Data de Vencimento | |––––|———————––|——|––––––––––| | 0 | 0.0 | 0.0 | 0.0 | | 1 | 2.0 | 3.0 | 5.0 | | 2 | 4.0 | 1.0 | 3.0 | | 3 | 3.0 | 4.0 | 6.0 | | 4 | 5.0 | 2.0 | 4.0 | | 5 | 1.0 | 5.0 | 7.0 |

pub(crate) fn load_from_file(filename: &str) -> Self

Função para carregar uma instância do problema a partir de um arquivo

Arguments

filename - Nome do arquivo contendo a instância do problema

Returns

Self - Nova instância do problema carregada do arquivo